+++ preview/preview.mm
@@ -18,7 +18,7 @@ int main(int argc, const char *argv[]) {
         candidateWindow->set_page_callback([](bool next) {
             std::cout << (next ? "next" : "prev") << " page" << std::endl;
         });
-        candidateWindow->set_action_callback([](size_t index, int id) {
+        candidateWindow->set_action_callback([](int index, int id) {
             std::cout << "action " << id << " on " << index << std::endl;
         });
         auto t = std::thread([&] {
